Clean project and add md
