+++ components/index.js
@@ -1,0 +1,4 @@
+export { default as Summary } from './topic/summary';
+export { default as Progress } from './topic/progress';
+export { default as Related } from './topic/related';
+export { default as References } from './topic/references';
