+++ components/layout.js
@@ -1,5 +1,5 @@
 import Head from 'next/head';
-import ButtonAppBar from './app-bar';
+import CustomAppBar from './app-bar';
 
 const Layout = ({ children }) => {
 	return (
@@ -11,7 +11,7 @@ const Layout = ({ children }) => {
 				<link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto:300,400,500,700&display=swap" />
 				<link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons" />
 			</Head>
-			<ButtonAppBar />
+			<CustomAppBar />
 			{children}
 		</div>
 	)
