+++ components/tabs/index.js
@@ -6,7 +6,7 @@ import Tabs from '@material-ui/core/Tabs';
 import Tab from '@material-ui/core/Tab';
 import Typography from '@material-ui/core/Typography';
 import Box from '@material-ui/core/Box';
-import Card from './card';
+import Card from '../card';
 
 const useStyles = makeStyles(theme => ({
 	root: {
