+++ components/topic/progress.js
@@ -1,0 +1,6 @@
+
+const Progress = () => (
+	<div></div>
+)
+
+export default Progress;
