+++ components/topic/summary.js
@@ -1,0 +1,7 @@
+const Summary = ({ content }) => (
+	<div>
+		<p>{content}</p>
+	</div>
+)
+
+export default Summary;
