+++ now.json
@@ -1,0 +1,7 @@
+{
+	"build": {
+		"env": {
+			"API": "http://3.130.98.8/"
+		}
+	}
+}
