+++ pages/index.js
@@ -3,20 +3,20 @@ import Layout from '../components/layout';
 import List from '../components/list';
 
 const Home = ({ data = [] }) => {
-  return (
+	return (
 		<Layout>
 			<List data={data} />
 		</Layout>
 	);
 }
 
-Home.getInitialProps = async ({ req }) => {
+Home.getInitialProps = async () => {
 	const res = await fetch(process.env.API + 'topics?date=')
 	try {
 		const json = await res.json()
 		console.log(json);
 		return { data: json }
-	} catch(e) {
+	} catch (e) {
 		return { data: null }
 	}
 }
