+++ readme.md
@@ -1,0 +1,16 @@
+### Little Gas Book
+
+The application is using basic next.js configuration.
+
+Before you start
+1) Put API link to `.env`
+2) `yarn install`
+
+.env sample
+```
+API=http://127.0.0.1/
+```
+
+Local develoment
+1) yarn run dev
+2) Go to http://localhost:3000 or the port you've set
